requirements.txt
alive-progress==1.6.2
attrs==21.2.0
awsebcli==3.20.2
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
blessed==1.19.0
boto3==1.18.53
botocore==1.21.53
cached-property==1.5.2
cement==2.8.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.3
convertapi==1.4.0
cryptography==35.0.0
distlib==0.3.3
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==7.13.4
filelock==3.4.0
Flask==2.0.2
future==0.16.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
paramiko==2.8.0
pathspec==0.5.9
platformdirs==2.4.0
pycparser==2.21
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
requests-aws4auth==1.1.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
semantic-version==2.8.5
six==1.14.0
termcolor==1.1.0
texttable==1.6.4
tqdm==4.62.2
Unidecode==1.2.0
urllib3==1.26.6
virtualenv==20.10.0
wcwidth==0.1.9
websocket-client==0.59.0
Werkzeug==2.0.2
